+++ R/DiseasystoreBase.R
@@ -281,11 +281,7 @@ DiseasystoreBase <- R6::R6Class(                                                
 
       coll <- checkmate::makeAssertCollection()
       checkmate::assert_choice(observable, self$available_observables, add = coll)
-      checkmate::assert(
-        checkmate::check_list(stratification, types = "character", null.ok = TRUE),
-        checkmate::check_multi_class(stratification, c("character", "quosure", "quosures"), null.ok = TRUE),
-        add = coll
-      )
+      checkmate::assert_multi_class(stratification, c("quosure", "quosures"), null.ok = TRUE, add = coll)
       checkmate::assert_date(start_date, add = coll)
       checkmate::assert_date(end_date, add = coll)
       checkmate::reportAssertions(coll)
@@ -351,14 +347,6 @@ DiseasystoreBase <- R6::R6Class(                                                
           unlist() |>
           unique()
 
-        # Report if stratification not found
-        if (is.null(stratification_features)) {
-          err <- glue::glue("Stratification variable not found. ",
-                            "Available stratification variables are: ",
-                            "{toString(self$available_stratifications)}")
-          stop(err, call. = FALSE)
-        }
-
         # Determine the name of the columns created by the stratifications
         stratification_names <- stratification |>
           purrr::map(rlang::as_label) |>
@@ -402,6 +390,7 @@ DiseasystoreBase <- R6::R6Class(                                                
       out <- truncate_interlace(observable_data, stratification_data) |>
         private$key_join_filter(stratification_features, start_date, end_date) |>
         dplyr::compute()
+      SCDB::defer_db_cleanup(out)
 
       # Retrieve the aggregators (and ensure they work together)
       key_join_aggregators <- c(purrr::pluck(private, purrr::pluck(ds_map, observable)) %.% key_join,
@@ -417,31 +406,46 @@ DiseasystoreBase <- R6::R6Class(                                                
 
       key_join_aggregator <- purrr::pluck(key_join_aggregators, 1)
 
+      # Check stratification can be performed:
+      out <- tryCatch(
+        dplyr::group_by(out, !!!stratification),
+        error = function(e) {
+          error_msg <- glue::glue(
+            "Stratification could not be computed. ",
+            "Error {tolower(substr(e$message, 1, 1))}{substr(e$message, 2, nchar(e$message))}. ",
+            "Available stratification variables are: ",
+            "{toString(self$available_stratifications)}"
+          )
+          stop(error_msg, call. = FALSE)
+        }
+      )
+
       # Add the new valid counts
       t_add <- out |>
-        dplyr::group_by(!!!stratification) |>
         dplyr::group_by("date" = .data$valid_from, .add = TRUE) |>
         key_join_aggregator(observable) |>
         dplyr::rename("n_add" = "n") |>
         dplyr::compute()
+      SCDB::defer_db_cleanup(t_add)
 
       # Add the new invalid counts
       t_remove <- out |>
-        dplyr::group_by(!!!stratification) |>
         dplyr::group_by("date" = .data$valid_until, .add = TRUE) |>
         key_join_aggregator(observable) |>
         dplyr::rename("n_remove" = "n") |>
         dplyr::compute()
+      SCDB::defer_db_cleanup(t_remove)
 
       # Get all combinations to merge onto
       all_dates <- tibble::tibble(date = seq.Date(from = start_date, to = end_date, by = 1))
 
       if (!is.null(stratification)) {
         all_combinations <- out |>
-          dplyr::ungroup() |>
-          dplyr::distinct(!!!stratification) |>
+          dplyr::select(dplyr::all_of(stratification_names)) |>
+          dplyr::distinct() |>
           dplyr::cross_join(all_dates, copy = TRUE) |>
           dplyr::compute()
+        SCDB::defer_db_cleanup(all_combinations)
       } else {
         all_combinations <- all_dates
 
@@ -469,12 +473,6 @@ DiseasystoreBase <- R6::R6Class(                                                
       # Ensure date is of type Date
       data <- data |>
         dplyr::mutate("date" = as.Date(.data$date))
-
-
-      # Clean up
-      DBI::dbRemoveTable(self %.% target_conn, SCDB::id(out))
-      DBI::dbRemoveTable(self %.% target_conn, SCDB::id(t_add))
-      DBI::dbRemoveTable(self %.% target_conn, SCDB::id(t_remove))
 
       return(data)
     }