+++ NEWS.md
@@ -1,5 +1,14 @@
 # diseasystore (development version)
 
+## Breaking change
+
+* `DiseasystoreBase$key_join_features()` no longer accept character stratifications -- must use `rlang::quos()` (#203).
+
+## Minor Improvements and Fixes
+
+* In `DiseasystoreBase$key_join_features()`, stratifications no longer need to do computation on other features (#203).
+
+
 # diseasystore 0.3.0
 
 ## Breaking change